--- src/constants/Page.js
@@ -1,3 +1,0 @@
-export const GET_PHOTOS_REQUEST = 'GET_PHOTOS_REQUEST'
-export const GET_PHOTOS_SUCCESS = 'GET_PHOTOS_SUCCESS'
-export const GET_PHOTOS_FAIL = 'GET_PHOTOS_FAIL'
