--- src/constants/User.js
@@ -1,3 +1,0 @@
-export const LOGIN_REQUEST = 'LOGIN_REQUEST'
-export const LOGIN_SUCCES = 'LOGIN_SUCCES'
-export const LOGIN_FAIL = 'LOGIN_FAIL'
