+++ src/modules/page.js
@@ -1,9 +1,35 @@
-import {
-  GET_PHOTOS_REQUEST,
-  GET_PHOTOS_FAIL,
-  GET_PHOTOS_SUCCESS
-} from '../constants/Page'
+//constants
+export const GET_PHOTOS_REQUEST = 'GET_PHOTOS_REQUEST'
+export const GET_PHOTOS_SUCCESS = 'GET_PHOTOS_SUCCESS'
+export const GET_PHOTOS_FAIL = 'GET_PHOTOS_FAIL'
 
+const initialState = {
+  year: 2016,
+  photos: [],
+  fetching: false,
+  error: ''
+}
+
+//reducer
+export default function page(state = initialState, action) {
+
+  switch (action.type) {
+    case GET_PHOTOS_REQUEST:
+      return { ...state, year: action.payload, fetching: true, error: '' }
+
+    case GET_PHOTOS_SUCCESS:
+      return { ...state, photos: action.payload, fetching: false, error: '' }
+
+    case GET_PHOTOS_FAIL:
+      return { ...state, error: action.payload.message, fetching: false }
+
+    default:
+      return state;
+  }
+
+}
+
+//actions
 let photosArr = []
 let cached = false
