+++ src/modules/reducer.js
@@ -1,4 +1,3 @@
-
 import { combineReducers } from 'redux'
 import page from './page'
 import user from './user'
