+++ src/store/configureStore.js
@@ -1,5 +1,5 @@
 import { createStore, applyMiddleware } from 'redux'
-import rootReducer from '../reducers'
+import rootReducer from '../modules/reducer'
 import createLogger from 'redux-logger'
 import thunk from 'redux-thunk' // <-- добавили redux-thunk
 
@@ -12,8 +12,8 @@ export default function configureStore(initialState) {
     applyMiddleware(thunk, logger)) // <-- добавили его в цепочку перед logger'ом
 
   if (module.hot) {
-    module.hot.accept('../reducers', () => {
-      const nextRootReducer = require('../reducers')
+    module.hot.accept('../modules/reducer', () => {
+      const nextRootReducer = require('../modules/reducer')
       store.replaceReducer(nextRootReducer)
     })
   }
